Title: 숫자 사이의 배수, Time: 6ms, Memory: 0MB, Status: Passed - CodeTree

+++ 241104/숫자 사이의 배수/multiplication-between-numbers.c
@@ -9,6 +9,6 @@ int main() {
             sum += i;
         }
     }
-    printf("%d %.1lf", sum, sum / cnt);
+    printf("%d %.1lf", sum, (double)sum / (double)cnt);
     return 0;
 }
